requirements.txt
django
django-dotenv
djangorestframework
django-filter
pillow
django-silk
drf-spectacular